requirements.txt
dash
plotly
psycopg2-binary==2.9.9
pandas==2.1.3
setuptools
wheel 
build